+++ Sources/RecogLib_iOS.xcframework/ios-arm64_x86_64-maccatalyst/RecogLib_iOS.framework/Versions/A/Headers/CImageSignature.hpp
@@ -1,0 +1,14 @@
+#ifndef CImageSignature_hpp
+#define CImageSignature_hpp
+
+#include <stdio.h>
+#include <CoreMedia/CoreMedia.h>
+
+struct CImageSignature {
+    const uint8_t *image;
+    int imageSize;
+    const char *signature;
+    int signatureSize;
+};
+
+#endif /* CImageSignature_hpp */