+++ Sources/RecogLib_iOS.xcframework/ios-arm64_x86_64-maccatalyst/RecogLib_iOS.framework/Versions/A/Headers/RecogLib_iOS.h
@@ -1,0 +1,15 @@
+#import <UIKit/UIKit.h>
+#import <RecogLib_iOS/CZenidSecurityWrapper.hpp>
+#import <RecogLib_iOS/CDocumentWrapper.hpp>
+#import <RecogLib_iOS/CFaceLivenessWrapper.hpp>
+#import <RecogLib_iOS/CSelfieWrapper.hpp>
+#import <RecogLib_iOS/CDocumentVerifierSettingsWrapper.hpp>
+#import <RecogLib_iOS/CZenidLogWrapper.hpp>
+
+//! Project version number for RecogLib_iOS.
+FOUNDATION_EXPORT double RecogLib_iOSVersionNumber;
+
+//! Project version string for RecogLib_iOS.
+FOUNDATION_EXPORT const unsigned char RecogLib_iOSVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <RecogLib_iOS/PublicHeader.h>