+++ CHANGELOG.md
@@ -1,8 +1,30 @@
 # Change Log
+## 2.0.9 2022-07-25
+* Updated LibZenid 2.8.3
+* New: Support for Dutch driving license (v2013 and v2014). ZenID-issues#2514
+* New: Support for Swiss passport (v2010). ZenID-issues#2514
+* Added Al, By, Ba, Me, Mk, Rs, Ch to Country.
+* Added AL_PAS_2009, BA_PAS_2014, CH_PAS_2010, ME_PAS_2008, MK_PAS_2007, RS_PAS_2008, NO_PAS_2020, NO_PAS_2011_15 in DocumentCodes enum.
+* Support for British passports (v2010/2011/2015/2019).
+* Fix of the issue "The smile animation is not shown"
+* New model for German ID also able to read 2021
+
+## 2.0.8 2022-07-25
+* Updated LibZenid 2.7.4
+* Fix Support for Czech birth certificate - version 2001. ZenID-issues#2500
+* Added usurf_hr_idc_2003_f.bin
+
+## 2.0.7 2022-07-18
+* Updated LibZenid 2.7.3
+* Fixed failed assertion in face liveness and selfie liveness introduced in 2.6.4. (websdk as well as sdks)
+* New: Support for Czech birth certificate - version 2001. ZenID-issues#2500
+* Added usurf_hr_idc_2003_f.bin
+* Add support for mac catalyst
 
 ## 2.0.6 2022-07-01
 * Updated LibZenid 2.6.3
 * Support for ARM simulator builds in iOS
+* Fix Faceliveness visualizer
 
 ## 2.0.5 2022-05-25
 * Updated LibZenid 2.5.1